refactor(package): draw_numbers/builder 패키지 구성
