+++ src/main/java/lotto/LottoController.java
@@ -4,7 +4,7 @@ import lotto.model.lotto.Lotto;
 import lotto.model.lotto.LottoChecker;
 import lotto.model.lotto.LottoPublisher;
 import lotto.model.draw_numbers.DrawNumbers;
-import lotto.model.draw_numbers.DrawNumbersBuilder;
+import lotto.model.draw_numbers.builder.DrawNumbersBuilder;
 import lotto.model.number_generator.LottoNumberGenerator;
 import lotto.model.number_generator.NumberGenerator;
 import lotto.view.InputView;
